Create Advertiser class. Major changes to AdvertiserUtil.
